add-host-sync-project-file Enables using host file sharing
By adding configurable host file syncing via the default file sharing
this modification provides an alternative to the guest-to-host NFS
sharing that this project has used to date.

This still is not as clean a solution, as the FS events are lost,
and the NFS share cannot overlap with a directory shared from the
host. This means that to have full synchronization rsync or some
other solution would still be needed to provide both speed of updates
and durable access on the host.

In practice runs it was taking 2-4s for files to update from the
host to the guest. I explored using Vagrant's rsync and rsync-auto
and those seem like brittle solutions.

My current thinking is that using a trigger to rsync targetted
directories prior to the halt would enable backup of files to the host
in a useful fashion. It could also be triggered after up.